14.12/01/django_aut_samp-main/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.4
tzdata==2025.2
